Update get-started.md
For iOS:
Cocoapods has removed the use of use_frameworks! and this breaks the react-native libraries. Adding the frameworks build phase, adding the .framework file to the top of the project and linking the libraries in XCode is no longer needed.


For Android:
sqipVersion = "1.1.0" doesn't work with the testing cards so I recommend a change to "1.2.0" which is the latest and works with the testing cards.